pyproject.toml
[project]
name = "hfm"
version = "0.2.0"
description = "Variable time molecular dynamics via hamiltonian flow maps"
readme = "README.md"
requires-python = "==3.12.9"
dependencies = [
    "nvitop",
    "python-dotenv",
    "ase>=3.25.0",
    "e3x>=1.0.2",
    "flax>=0.10.7",
    "jraph>=0.0.6.dev0",
    "nglview>=3.1.4,<4",
    "numpy<2.3.0",
    "pip>=25.1.1",
    "tqdm>=4.67.1",
    "wandb>=0.20.1",
    "hydra-core>=1.3.2",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "cloudpickle==2.2.1",
    "submitit==1.5.2",
    "hydra-submitit-launcher@git+ssh://git@github.com/winfried-ripken/hydra.git@main#subdirectory=plugins/hydra_submitit_launcher",
    "seaborn>=0.13.2",
    "pandas>=2.3.3",
    "moviepy==0.2.2.11",
]

[project.optional-dependencies]
cuda12 = [
    "jax[cuda12]==0.8.1",
    "jaxlib==0.8.1",
    "jax-cuda12-pjrt==0.8.1",
    "jax-cuda12-plugin==0.8.1",
]
cuda13 = [
    "jax[cuda13]==0.8.1",
    "jaxlib==0.8.1",
    "jax-cuda13-pjrt==0.8.1",
    "jax-cuda13-plugin==0.8.1",
]
tfds = [
    "tensorflow-cpu>=2.20.0",
    "tensorflow-datasets>=4.9.9",
    "apache-beam>=2.69.0",
    "mlcroissant>=1.0.22",
    "h5py>=3.15.1",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
exclude = [
    "build*",
    "configs*",
    "data*",
    "examples*",
    "examples_deprecated*",
    "old_model*",
    "tests*",
    "tf_datasets*",
    "toy_examples*",
    "hfm.multirun*",
    "hfm.outputs*",
]
